fix: add cursor party to generative layout
